Speed up realtime tracking: detect-throttling + opt-in CUDA (#13)
Profiled the ~0.5 FPS pipeline: S3FD detect ~1.27s/frame (dominant),
FAN forward ~0.6s/frame, capture negligible. Root cause: candle-core/
candle-nn build with no acceleration backend by default, so every
forward pass runs candle's CPU fallback even on machines with a GPU.

- FanTracker now only re-runs the (expensive) detector every
  `detect_interval` frames (default 8, --detect-interval N); frames in
  between derive their crop box from the previous frame's FAN
  landmarks. Always re-detects immediately after losing the face.
  TDD'd with a stub FaceDetector call-counter.
- Add an opt-in, non-default `cuda` Cargo feature (candle-core/cuda +
  candle-nn/cuda) and switch Device::Cpu -> Device::cuda_if_available(0)
  in both FanTracker/SfdDetector loaders, which no-ops back to CPU
  when the feature is off or no GPU is found. Not part of `default`:
  macOS has no CUDA and most end users won't have an NVIDIA GPU + CUDA
  toolkit to build against.

Measured on this machine (16-core Ryzen, RTX 5090): detect-throttling
alone roughly doubles CPU throughput; `--features cuda` reaches ~29.5
FPS (capture-rate-limited) vs ~0.5 FPS CPU-only baseline.

+++ src/tracking/fan/preprocess.rs
@@ -50,8 +50,8 @@ pub fn crop(image: &RgbImage, center: (f32, f32), scale: f32, resolution: u32) -
     let ul = transform((1.0, 1.0), center, scale, res_f, true);
     let br = transform((res_f, res_f), center, scale, res_f, true);
 
-    let new_w = (br.0 - ul.0) as i64;
-    let new_h = (br.1 - ul.1) as i64;
+    let new_w = br.0 - ul.0;
+    let new_h = br.1 - ul.1;
     let (img_w, img_h) = (image.width() as i64, image.height() as i64);
 
     let mut canvas = RgbImage::new(new_w.max(1) as u32, new_h.max(1) as u32);
